+++ resources/views/admin/operatingsystem/create.blade.php
@@ -1,5 +1,5 @@
 <x-admin-layout>
-    <x-create.main :header="__('Neues Betriebsystem')" action="{{ route('admin.operatingsystem.store') }}">
+    <x-create.main :header="__('Neues Betriebssystem')" action="{{ route('admin.operatingsystem.store') }}">
 
         <x-create.singlerow :label="__('Name')" name="name" />