+++ resources/views/admin/operatingsystem/edit.blade.php
@@ -1,5 +1,5 @@
 <x-admin-layout>
-    <x-create.main :header="__('Betriebsystem bearbeiten')" :labelsubmit="__('Speichern')" action="{{ route('admin.operatingsystem.update', $operatingSystem) }}">
+    <x-create.main :header="__('Betriebssystem bearbeiten')" :labelsubmit="__('Speichern')" action="{{ route('admin.operatingsystem.update', $operatingSystem) }}">
         @method('PATCH')
 
         <x-create.singlerow :label="__('Name')" name="name" :default="$operatingSystem->name" />